feat: add error_handler for commit-check
